+++ apps/server/src/apps/server.app.ts
@@ -95,7 +95,7 @@ async function bootstrap() {
 	// eslint-disable-next-line @typescript-eslint/no-unsafe-argument
 	rootExpress.use('/', logDeprecatedPaths, feathersExpress);
 
-	const port = 3030;
+	const port = 3031;
 	rootExpress.listen(port);
 
 	console.log('#################################');
