+++ apps/server/src/modules/tool/tool-api.module.ts
@@ -3,6 +3,7 @@ import { LtiToolRepo } from '@shared/repo';
 import { LoggerModule } from '@src/core/logger';
 import { AuthorizationModule } from '@src/modules/authorization';
 import { UserModule } from '@src/modules/user';
+import { SchoolModule } from '@src/modules/school';
 import { ExternalToolRequestMapper, ExternalToolResponseMapper, Lti11ResponseMapper } from './controller/mapper';
 import { ToolConfigurationController } from './controller/tool-configuration.controller';
 import { ToolController } from './controller/tool.controller';
@@ -10,11 +11,10 @@ import { ToolModule } from './tool.module';
 import { ExternalToolUc } from './uc/external-tool.uc';
 import { Lti11Uc } from './uc/lti11.uc';
 import { LtiRoleMapper } from './uc/mapper';
-import { ExternalToolConfigurationUc } from './uc/tool-configuration.uc';
+import { ExternalToolConfigurationUc } from './uc/external-tool-configuration.uc';
 import { ToolSchoolController } from './controller/tool-school.controller';
 import { SchoolExternalToolResponseMapper } from './controller/mapper/school-external-tool-response.mapper';
 import { SchoolExternalToolUc } from './uc/school-external-tool.uc';
-import { SchoolModule } from '@src/modules/school';
 
 @Module({
 	imports: [ToolModule, UserModule, AuthorizationModule, LoggerModule, SchoolModule],
