+++ apps/server/src/modules/tool/uc/external-tool-configuration.uc.ts
@@ -2,16 +2,37 @@ import { Injectable } from '@nestjs/common';
 import { Actions, CustomParameterScope, EntityId, Permission } from '@shared/domain';
 import { ExternalToolDO } from '@shared/domain/domainobject/external-tool';
 import { NotFoundException } from '@nestjs/common/exceptions/not-found.exception';
+import { AuthorizationService } from '@src/modules/authorization';
+import { AllowedAuthorizationEntityType } from '@src/modules/authorization/interfaces';
+import { Page } from '@shared/domain/interface/page';
+import { SchoolExternalToolDO } from '@shared/domain/domainobject/external-tool/school-external-tool.do';
 import { ExternalToolService } from '../service/external-tool.service';
-import { AuthorizationService } from '../../authorization';
-import { AllowedAuthorizationEntityType } from '../../authorization/interfaces';
+import { SchoolExternalToolService } from '../service/school-external-tool.service';
 
 @Injectable()
 export class ExternalToolConfigurationUc {
 	constructor(
 		private readonly externalToolService: ExternalToolService,
+		private readonly schoolExternalToolService: SchoolExternalToolService,
 		private readonly authorizationService: AuthorizationService
 	) {}
+
+	async getAvailableToolsForSchool(userId: EntityId, schoolId: EntityId): Promise<ExternalToolDO[]> {
+		await this.ensureSchoolPermission(userId, schoolId);
+
+		const externalTools: Page<ExternalToolDO> = await this.externalToolService.findExternalTools({});
+		const toolsInUse: SchoolExternalToolDO[] = await this.schoolExternalToolService.findSchoolExternalTools({
+			schoolId,
+		});
+		const toolIdsInUse: EntityId[] = toolsInUse.map(
+			(schoolExternalTool: SchoolExternalToolDO): EntityId => schoolExternalTool.toolId
+		);
+
+		const availableTools: ExternalToolDO[] = externalTools.data.filter(
+			(tool: ExternalToolDO): boolean => !tool.isHidden && !!tool.id && !toolIdsInUse.includes(tool.id)
+		);
+		return availableTools;
+	}
 
 	async getExternalToolForSchool(
 		userId: EntityId,
