--- apps/server/src/modules/tool/uc/tool-configuration.uc.ts
@@ -1,47 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { Actions, EntityId, Permission } from '@shared/domain';
-import { ExternalToolDO } from '@shared/domain/domainobject/external-tool';
-import { SchoolExternalToolDO } from '@shared/domain/domainobject/external-tool/school-external-tool.do';
-import { Page } from '@shared/domain/interface/page';
-import { AuthorizationService } from '@src/modules/authorization';
-import { AllowedAuthorizationEntityType } from '@src/modules/authorization/interfaces';
-import { ExternalToolService } from '../service/external-tool.service';
-import { SchoolExternalToolService } from '../service/school-external-tool.service';
-
-@Injectable()
-export class ExternalToolConfigurationUc {
-	constructor(
-		private readonly externalToolService: ExternalToolService,
-		private readonly schoolExternalToolService: SchoolExternalToolService,
-		private readonly authorizationService: AuthorizationService
-	) {}
-
-	async getAvailableToolsForSchool(userId: EntityId, schoolId: EntityId): Promise<ExternalToolDO[]> {
-		await this.ensureSchoolPermission(userId, schoolId);
-
-		const externalTools: Page<ExternalToolDO> = await this.externalToolService.findExternalTools({});
-		const toolsInUse: SchoolExternalToolDO[] = await this.schoolExternalToolService.findSchoolExternalTools({
-			schoolId,
-		});
-		const toolIdsInUse: EntityId[] = toolsInUse.map(
-			(schoolExternalTool: SchoolExternalToolDO): EntityId => schoolExternalTool.toolId
-		);
-
-		const availableTools: ExternalToolDO[] = externalTools.data.filter(
-			(tool: ExternalToolDO): boolean => !tool.isHidden && !!tool.id && !toolIdsInUse.includes(tool.id)
-		);
-		return availableTools;
-	}
-
-	private async ensureSchoolPermission(userId: EntityId, schoolId: EntityId) {
-		return this.authorizationService.checkPermissionByReferences(
-			userId,
-			AllowedAuthorizationEntityType.School,
-			schoolId,
-			{
-				action: Actions.read,
-				requiredPermissions: [Permission.SCHOOL_TOOL_ADMIN],
-			}
-		);
-	}
-}
